+++ bench_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgconn/stmtcache"
 	"github.com/khaibin/pgtype"
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 	"github.com/stretchr/testify/require"
 )