+++ log/logrusadapter/adapter.go
@@ -5,7 +5,7 @@ package logrusadapter
 import (
 	"context"
 
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 	"github.com/sirupsen/logrus"
 )