+++ pgxpool/conn.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/jackc/pgconn"
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 	"github.com/jackc/puddle"
 )