+++ pgxpool/conn_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/jackc/pgx/v4/pgxpool"
+	"github.com/khaibin/pgx/v4/pgxpool"
 	"github.com/stretchr/testify/require"
 )