+++ pgxpool/pool.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/jackc/pgconn"
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 	"github.com/jackc/puddle"
 	errors "golang.org/x/xerrors"
 )