+++ stdlib/sql.go
@@ -65,7 +65,7 @@ import (
 
 	"github.com/jackc/pgconn"
 	"github.com/khaibin/pgtype"
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 )
 
 // Only intrinsic types should be binary format with database/sql.