+++ tx_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/jackc/pgconn"
-	"github.com/jackc/pgx/v4"
+	"github.com/khaibin/pgx/v4"
 	"github.com/stretchr/testify/require"
 )